+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [[0.0.0-alpha.6](https://github.com/multiversx/mx-sdk-dapp-core-ui/pull/53)] - 2025-03-12
+
+- [Added notifications feed side panel](https://github.com/multiversx/mx-sdk-dapp-core-ui/pull/50)
+- [Added Tailwind CSS support](https://github.com/multiversx/mx-sdk-dapp-core/pull/47)
+- [Added wallet connect side panel component](https://github.com/multiversx/mx-sdk-dapp-core/pull/45)
+
 ## [[0.0.0-alpha.5](https://github.com/multiversx/mx-sdk-dapp-core-ui/pull/49)] - 2025-02-27
 
 - [Added ability to expand/collapse transactions in toast](https://github.com/multiversx/mx-sdk-dapp-core/pull/48)